api/requirements.txt
aniso8601==9.0.1
bcrypt==3.2.2
blinker==1.5
cffi==1.15.1
click==8.1.3
flask-cors==3.0.10
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.17.0
packaging==21.3
psycopg-binary==3.1.4
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.2.1
zipp==3.8.1
